requirements.txt
numpy==1.18.1
scipy==1.4.1
pandas==1.0.1
scikit-learn==0.22.1
tensorflow-gpu == 2.0.0
Cython==0.29.13
